simple_tavily_adapter/pyproject.toml
[project]
name = "searxng-tavily-adapter"
version = "1.0.0"
description = "SearXNG Tavily API adapter with Markdown content support and HTML sanitization"
authors = [{name = "SearXNG Tavily Adapter Team"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "aiohttp>=3.9.1",
    "pydantic>=2.5.1",
    "beautifulsoup4>=4.12.2",
    "pyyaml>=6.0.1",
    "markdownify>=1.2.0",
    "nh3>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.scripts]
tavily-adapter = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
